requirements.txt
absl-py==2.0.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dlib==19.24.2
flatbuffers==23.5.26
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
gast==0.2.2
google-auth==2.23.2
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637578017926/work
h5py @ file:///tmp/abs_4aewd3wzey/croots/recipe/h5py_1659091371897/work
idna==3.4
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
imutils==0.5.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
libclang==16.0.6
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe==2.1.3
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
munkres==1.1.4
numpy==1.17.5
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1671697413597/work
Pillow==9.4.0
ply==3.11
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3==2.0.5
webencodings==0.5.1
Werkzeug==2.2.3
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
zipp @ file:///croot/zipp_1672387121353/work
